2/4 is finished (?)

+++ scrapper/src/main/java/io/swagger/api/JdbcLinkRepository.java
@@ -14,6 +14,7 @@ import java.sql.Timestamp;
 import java.time.OffsetTime;
 import java.util.ArrayList;
 import java.util.Collection;
+import java.util.List;
 
 @Repository
 public class JdbcLinkRepository {
@@ -26,11 +27,13 @@ public class JdbcLinkRepository {
 
     public void save(Long id, String link) throws ApiException {
         try {
-            String query = ("select * from id where id=") + id.toString();
-            jdbcTemplate.queryForObject(query, Long.class);
+            String query = ("select * from id where id=?");
+            var first = jdbcTemplate.queryForList(query,id).toArray();
+            if (first.length == 0) throw new ApiException(404, "Чат не существует");
             var time = new Timestamp(System.currentTimeMillis());
             MapSqlParameterSource params = new MapSqlParameterSource();
-
+            var check = jdbcTemplate.queryForList("select id from connect where link=? and id=?", link,id).toArray();
+            if (check.length > 0 ) throw new ApiException(400,"Ссылка уже добавлена");
             params.addValue("id",id);
             params.addValue("link",link);
             params.addValue("updated",time);
@@ -43,30 +46,29 @@ public class JdbcLinkRepository {
 
     public void remove(Long id, String link)throws ApiException {
         try {
-            String query = ("select * from id where id=") + id.toString();
-            jdbcTemplate.queryForObject(query, Long.class);
+            String query = ("select * from id where id=?");
+            var doo = jdbcTemplate.queryForList(query, id).toArray();
+            if (doo.length == 0) throw new ApiException(404,"Чат не существует");
+            var check = jdbcTemplate.queryForList("select id from connect where link=? and id=?", link,id).toArray();
+            if (check.length == 0 ) throw new ApiException(400,"Ссылка не отслеживается");
             jdbcTemplate.update("delete from connect where link =? and id = ?", link,id);
         } catch (EmptyResultDataAccessException e) {
             throw new ApiException(404,"Чат не существует");
         }
     }
 
-    public Collection<URI> findAll(Long id)throws ApiException {
+    public Collection<URI> findAll(Long id) throws ApiException {
         try {
-            String query = ("select * from id where id=") + id.toString();
-            jdbcTemplate.queryForObject(query, Long.class);
-            ResultSet res = jdbcTemplate.queryForObject("select link from connect where id="+id+";",ResultSet.class);
-            ArrayList<URI> result = new ArrayList<>();
-            if (res != null && res.last()) {
-                result = new ArrayList<>(res.getRow());
-                int i =0;
-                while (res.next()) {
-                    result.add(i,new URI(res.getObject("link",String.class)));
-                    i++;
-                }
+            String query = ("select * from id where id=?");
+            var doo = jdbcTemplate.queryForList(query, id).toArray();
+            if (doo.length == 0) throw new ApiException(404,"Чат не существует");
+            var res = jdbcTemplate.queryForList("select link from connect where id="+id+";",String.class).toArray();
+            var result = new URI[res.length];
+            for (int i =0; i < res.length;i++) {
+                result[i]=new URI(res[i].toString());
             }
-            return result;
-        } catch (EmptyResultDataAccessException | SQLException e) {
+            return List.of(result);
+        } catch (EmptyResultDataAccessException e) {
             throw new ApiException(404,"Чат не существует");
         } catch (URISyntaxException e) {
             throw new RuntimeException(e);
